requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-magic>=0.4.24
psycopg2-binary>=2.9.1
pgvector>=0.1.8
sentence-transformers>=2.2.2
nltk>=3.6.3
numpy>=1.21.0
pydantic>=1.8.2
python-dotenv>=0.19.0
aiofiles>=0.7.0
PyPDF2>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-cloud-aiplatform==1.36.4
google-cloud-storage==2.13.0
google-auth==2.23.4
google-api-core==2.15.0
google-genai>=1.0.0
requests==2.31.0 